Minor Var Type Update
Changed floats to doubles in Lab3_2.c

+++ york/Lab3_2.c
@@ -20,8 +20,8 @@ int main(int argc, char* argv[]) {
     int numPairs = 0;
     int* intArrX = NULL;
     int* intArrY = NULL;
-    float euclidDistance = 0;
-    float minDistance = 0;
+    double euclidDistance = 0;
+    double minDistance = 0;
     int closePoints[4];
 
     // Open file, assign to fileHandle
